Create index.php
Changes the index.html into a php file, this will check whether or not the user has been logged correctly. If the user has been logged in correct then, they will be brought to userSelection.php. If the user has not been logged in then, an error message will be prompted under the password input.